+++ src/run/mod.rs
@@ -4,6 +4,7 @@
 #![doc = crate::_DOC_RUN_MODULES!()]
 #![doc = crate::_doc!(flat:"run")]
 #![doc = crate::_doc!(newline)]
+#![doc = crate::_QUOTE_RUN!()]
 //!
 //
 // safety
@@ -17,7 +18,7 @@ crate::CONST! { pub(crate) _DOC_RUN_MODULES =
 mod tick; // RuntimeTick
 
 pub mod frame; // FramePacer // WIP
-// pub mod setup; // RunCap* WIP
+pub mod setup; // RunCap* WIP
 
 crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _mods {
@@ -29,7 +30,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _pub_mods {
         pub use super::{
             frame::_all::*,
-            // setup::_all::*,
+            setup::_all::*,
         };
     }
     _crate_internals {